Refactor monitor from window code into extension method.
