+++ src/MrCapitalQ.FollowAlong.Core/Monitors/MonitorExtensions.cs
@@ -1,3 +1,4 @@
+using Microsoft.UI.Xaml;
 using System;
 using System.Runtime.InteropServices;
 using Windows.Graphics.Capture;
@@ -16,6 +17,12 @@ namespace MrCapitalQ.FollowAlong.Core.Monitors
             Marshal.Release(itemPointer);
 
             return item;
+        }
+
+        public static MonitorInfo? GetWindowMonitorSize(this Window window)
+        {
+            var hwnd = WinRT.Interop.WindowNative.GetWindowHandle(window);
+            return MonitorInterops.GetMonitorFromWindow(hwnd);
         }
 
         [ComImport]
